+++ types/module.d.ts
@@ -1,0 +1,3 @@
+import type { KinetoModule } from './index.js';
+declare const module: KinetoModule;
+export default module;